+++ src/workbench/bordo.mjs
@@ -386,7 +386,12 @@ export default syntax => ({
     // ---------------------------
     // Minimap
     // ---------------------------
-    "minimap.findMatchHighlight": "#f18eb0ee"
+    "minimap.findMatchHighlight": "#f18eb0ee",
+    "minimap.errorHighlight": "#e34e1cee",
+    "minimap.warningHighlight": "#e69533ee",
+    "minimapGutter.addedBackground": "#16b673",
+    "minimapGutter.modifiedBackground": "#49e9a6",
+    "minimapGutter.deletedBackground": "#e34e1c"
   },
   tokenColors: syntax
 });